+++ src/elite/apps/EliteSharp/SDLProgram.cs
@@ -21,6 +21,7 @@ namespace EliteSharp;
 internal static class SDLProgram
 {
     private const string Title = "Elite - The Sharp Kind";
+    private const string AssetLogCategory = "Assets";
 
     // Get these from config
     ////#if QHD
@@ -93,9 +94,18 @@ internal static class SDLProgram
 
         ServiceCollection services = new();
         services.AddSingleton(loggerFactory);
-        services.AddSingleton<IAbstraction>(_ => graphicsBackend == GraphicsBackend.Hardware
-            ? new SDLAbstraction(ScreenWidth, ScreenHeight, Title, AssetLocator.Create())
-            : new SoftwareAbstraction(ScreenWidth, ScreenHeight, Title));
+        services.AddSingleton<IAbstraction>(sp => graphicsBackend == GraphicsBackend.Hardware
+            ? new SDLAbstraction(
+                ScreenWidth,
+                ScreenHeight,
+                Title,
+                AssetLocator.Create(),
+                sp.GetRequiredService<ILoggerFactory>().CreateLogger(AssetLogCategory))
+            : new SoftwareAbstraction(
+                ScreenWidth,
+                ScreenHeight,
+                Title,
+                sp.GetRequiredService<ILoggerFactory>().CreateLogger(AssetLogCategory)));
         services.AddSingleton(sp => sp.GetRequiredService<IAbstraction>().Graphics);
         services.AddSingleton(sp => sp.GetRequiredService<IAbstraction>().Sound);
         services.AddSingleton(sp => sp.GetRequiredService<IAbstraction>().Keyboard);